# openapi.yaml
openapi: 3.1.0
info:
  title: Pipeline Labs SDK API
  version: 1.0.0
  description: |
    Curated SDK surface for Stainless. This spec is intentionally limited to the
    most relevant programmatic routes for dataset ingestion, AI processing, job
    tracking, downloads, chat, auth verification, billing, and health.
servers:
  - url: https://pipeline-labs.vercel.app
    description: Production
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/datasets/{id}/preview:
    get:
      operationId: previewDataset
      tags: [Datasets]
      summary: Preview dataset rows and columns
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: rows
          schema: { type: integer, default: 10, maximum: 100 }
      responses:
        '200':
          description: Dataset preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetPreview'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/datasets/{id}/download:
    get:
      operationId: downloadDataset
      tags: [Datasets]
      summary: Download a source or processed dataset
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: format
          schema:
            type: string
            enum: [csv, json, xlsx]
            default: csv
        - in: query
          name: processed
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: File stream
  /api/process:
    post:
      operationId: createProcessingJob
      tags: [Processing]
      summary: Create a chunked AI dataset processing job
      description: |
        Generates a chunked cleaning plan, processes dataset chunks with AI,
        validates output against the original schema, stores the cleaned dataset,
        and returns a processing summary plus download URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessRequest'
      responses:
        '200':
          description: Processing completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      operationId: getProcessingJob
      tags: [Processing]
      summary: Get processing job status
      parameters:
        - in: query
          name: jobId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Processing job state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingJob'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    DatasetPreview:
      type: object
      properties:
        columns:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              type: { type: string }
              sample:
                type: array
                items: { type: string }
        rows:
          type: array
          items:
            type: object
        totalRows: { type: integer }
        totalColumns: { type: integer }
    ProcessRequest:
      type: object
      required: [datasetId, prompt]
      properties:
        datasetId: { type: string, format: uuid }
        prompt: { type: string, maxLength: 4000 }
        options:
          type: object
          properties:
            outputFormat:
              type: string
              enum: [csv, json]
              default: csv
    ProcessResponse:
      type: object
      properties:
        jobId: { type: string, format: uuid }
        datasetId: { type: string, format: uuid }
        processedDatasetId: { type: string, format: uuid }
        status: { type: string, enum: [completed] }
        downloadUrl: { type: string, format: uri }
        previewUrl: { type: string }
        summary:
          type: object
          properties:
            originalRows: { type: integer }
            originalColumns: { type: integer }
            processedRows: { type: integer }
            processedColumns: { type: integer }
            operationsApplied:
              type: array
              items: { type: string }
            llmInstructions: { type: string }
            processingDurationMs: { type: integer }
            chunkCount: { type: integer }
            chunkSize: { type: integer }
    ProcessingJob:
      type: object
      properties:
        jobId: { type: string, format: uuid }
        datasetId: { type: string, format: uuid }
        processedDatasetId:
          type: string
          format: uuid
          nullable: true
        status: { type: string }
        currentStep: { type: integer }
        totalSteps: { type: integer }
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        downloadUrl:
          type: string
          format: uri
          nullable: true
        previewData:
          type: object
          nullable: true
        plan:
          type: object
          nullable: true
        progressMessages:
          type: array
          items:
            type: object
        error:
          type: string
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }